Give each CP a unified instruction/data cache.
Each Central Processor has a 32KB unified instruction/data cache, and should not only consult it before the Central Memory, but also  interact with the CM in terms of entire cache lines instead of in terms of individual bytes and words, for performance.

Still to do: Add a way to have the Central Memory trigger eviction of written-to cache lines on other Central Processors.